+++ backend/configurations/app_configurations.py
@@ -12,6 +12,7 @@ OPENAI_API_KEY = os.getenv('OPENAI_API_KEY')
 azure_endpoint = os.getenv("AZURE_OPENAI_ENDPOINT")
 azure_openai_api_version = os.getenv("AZURE_OPENAI_API_VERSION")
 azure_openai_model = os.getenv("AZURE_OPENAI_DEPLOYMENT")
+MCP_SERVER_URL = os.getenv("MCP_SERVER_URL")
 
 LOG_FILE="./logs/resume_parser.log"
 os.makedirs(os.path.dirname(LOG_FILE), exist_ok=True)
