+++ backend/configurations/openai_client_config.py
@@ -1,5 +1,5 @@
 from openai import AzureOpenAI
-from backend.configurations.config import OPENAI_API_KEY, azure_endpoint, azure_openai_api_version, azure_openai_model, setup_logger
+from backend.configurations.app_configurations import OPENAI_API_KEY, azure_endpoint, azure_openai_api_version, azure_openai_model, setup_logger
 
 logger = setup_logger()
